Export of findPackageInfo
